+++ Useful/API.m
@@ -110,7 +110,7 @@ static NSString* uuidHeader;
         // HEADERS
         [httpData appendData:[[NSString stringWithFormat:@"Content-Disposition: %@\r\n", part.contentDisposition] dataUsingEncoding:NSUTF8StringEncoding]];
         [httpData appendData:[[NSString stringWithFormat:@"Content-Type: %@\r\n", part.contentType] dataUsingEncoding:NSUTF8StringEncoding]];
-        [httpData appendData:[[NSString stringWithFormat:@"Content-Length: %u\r\n", part.contentData.length] dataUsingEncoding:NSUTF8StringEncoding]];
+        [httpData appendData:[[NSString stringWithFormat:@"Content-Length: %lu\r\n", (unsigned long)part.contentData.length] dataUsingEncoding:NSUTF8StringEncoding]];
         // EMPTY
         [httpData appendData:[@"\r\n" dataUsingEncoding:NSUTF8StringEncoding]];
         // CONTENT + newline
