+++ Useful/FunTypes.h
@@ -58,7 +58,7 @@ void asyncMain(Block block);
 void asyncBackground(Block block);
 void vibrateDevice();
 NSString* concat(id arg1, ...);
-NSNumber* num(int i);
+NSNumber* num(long i);
 NSNumber* numf(float f);
 void repeat(NSUInteger times, NSUIntegerBlock block);
